Link to blog post for non-GH-confident volunteers
I wrote up a blog post with some help for folks who haven't used GitHub, or aren't super comfortable with it). If that's useful to have linked in the Readme, this PR adds a line under the "I am not confident" section linking to the post. No worries if it isn't useful!