Correcting English posts to use "Indigenous" re:#139
Partially addresses #139; still needed are any changes to posts in Spanish, and guidance in a visible place for language to use in future posts. I left the words in place in one post (Christopher Columbus—St. Paul, MN, USA) because it appears in the phrase "American Indian Movement", which looks like the formal name of a movement?